requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
faiss-cpu==1.7.3
fastapi==0.92.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
zipp==3.14.0
